ingestion/load_to_gcp_bucket/requirements.txt
google-cloud-storage==3.10.1
requests==2.33.1
dotenv==0.9.9